requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
html2text>=2025.4.15
python-dotenv
langdetect>=1.0.9
langgraph>=0.0.10
langchain-openai>=0.0.2
langchain-core>=0.1.0
